reorganize tests fixtures according to source
