+++ tests/fixtures/api/pj/musee_picasso.json
