+++ tests/fixtures/api/pj/musee_picasso_short.json
