+++ tests/fixtures/api/tripadvisor_hotel_pricing/__init__.py
