+++ tests/fixtures/api/tripadvisor_hotel_pricing/api_search_by_hotel_id_the_hotel_captain_cook.json
