+++ tests/fixtures/geocodeur/autocomplete/__init__.py
@@ -8,17 +8,17 @@ NLU_URL = "http://qwant.nlu/"
 CLASSIF_URL = "http://qwant.classif"
 ES_URL = "http://qwant.es"
 
-FIXTURE_AUTOCOMPLETE = read_fixture("fixtures/autocomplete/pavillon_paris.json")
+FIXTURE_AUTOCOMPLETE = read_fixture("fixtures/geocodeur/autocomplete/osm/pavillon_paris.json")
 
 FIXTURE_CLASSIF = {
-    "pharmacie": read_fixture("fixtures/autocomplete/classif_pharmacy.json"),
-    "bank": read_fixture("fixtures/autocomplete/classif_bank.json"),
+    "pharmacie": read_fixture("fixtures/nlp/classifier/classif_pharmacy.json"),
+    "bank": read_fixture("fixtures/nlp/classifier/classif_bank.json"),
 }
 
 
 def mock_NLU_for(httpx_mock, dataset):
     with override_settings({"NLU_TAGGER_URL": NLU_URL, "NLU_CLASSIFIER_URL": CLASSIF_URL}):
-        nlu_json = read_fixture(f"fixtures/autocomplete/nlu/{dataset}.json")
+        nlu_json = read_fixture(f"fixtures/nlp/nlu/{dataset}.json")
         httpx_mock.post(NLU_URL).respond(json=nlu_json)
 
         for q, data in FIXTURE_CLASSIF.items():
@@ -88,27 +88,29 @@ def mock_NLU_with_city(httpx_mock):
 def mock_autocomplete_get(httpx_mock):
     with override_settings({"BRAGI_BASE_URL": BASE_URL}):
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete.*q=bloublou")).respond(
-            json=read_fixture("fixtures/autocomplete/empty.json")
+            json=read_fixture("fixtures/geocodeur/autocomplete/osm/empty.json")
         )
 
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete.*q=(paris|parigi).*")).respond(
-            json=read_fixture("fixtures/autocomplete/paris.json")
+            json=read_fixture("fixtures/geocodeur/autocomplete/osm/paris.json")
         )
 
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete.*q=auchan.*")).respond(
-            json=read_fixture("fixtures/autocomplete/auchan.json")
+            json=read_fixture("fixtures/geocodeur/autocomplete/osm/auchan.json")
         )
 
         httpx_mock.get(
             re.compile(rf"^{BASE_URL}/autocomplete.*q=43\+rue\+de\+paris\+rennes.*")
-        ).respond(json=read_fixture("fixtures/autocomplete/43_rue_de_paris_rennes.json"))
+        ).respond(
+            json=read_fixture("fixtures/geocodeur/autocomplete/osm/43_rue_de_paris_rennes.json")
+        )
 
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete.*q=hotel.*poi_dataset.*")).respond(
-            json=read_fixture("fixtures/autocomplete/tripadvisor/hotel_moliere.json")
+            json=read_fixture("fixtures/geocodeur/autocomplete/tripadvisor/hotel_moliere.json")
         )
 
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete.*q=chez.*poi_dataset.*")).respond(
-            json=read_fixture("fixtures/autocomplete/tripadvisor/chez_eric.json")
+            json=read_fixture("fixtures/geocodeur/autocomplete/tripadvisor/chez_eric.json")
         )
 
         httpx_mock.get(re.compile(f"^{BASE_URL}/autocomplete")).respond(json=FIXTURE_AUTOCOMPLETE)
@@ -132,7 +134,7 @@ def mock_autocomplete_unavailable(httpx_mock):
 
 @pytest.fixture
 def mock_bragi_carrefour_in_bbox(request, httpx_mock):
-    bragi_response = read_fixture("fixtures/autocomplete/carrefour_in_bbox.json")
+    bragi_response = read_fixture("fixtures/geocodeur/autocomplete/osm/carrefour_in_bbox.json")
     limit = getattr(request, "param", {}).get("limit")
     if limit is not None:
         bragi_response["features"] = bragi_response["features"][:limit]
