+++ tests/fixtures/geocodeur/autocomplete/osm/43_rue_de_paris_rennes.json
