+++ tests/fixtures/geocodeur/autocomplete/osm/auchan.json
