+++ tests/fixtures/geocodeur/autocomplete/osm/carrefour_in_bbox.json
