+++ tests/fixtures/geocodeur/autocomplete/osm/empty.json
