+++ tests/fixtures/geocodeur/autocomplete/osm/pavillon_paris.json
