+++ tests/fixtures/geocodeur/autocomplete/tripadvisor/chez_eric.json
