+++ tests/fixtures/geocodeur/autocomplete/tripadvisor/hotel_moliere.json
