+++ tests/fixtures/geocodeur/search/osm/43_rue_de_paris_rennes.json
