+++ tests/fixtures/geocodeur/search/osm/auchan.json
