+++ tests/fixtures/geocodeur/search/osm/pavillon_paris.json
