+++ tests/fixtures/geocodeur/search/tripadvisor/hotel_moliere.json
