+++ tests/fixtures/nlp/classifier/classif_bank.json
