+++ tests/fixtures/nlp/classifier/classif_pharmacy.json
