+++ tests/fixtures/nlp/nlu/with_brand.json
