+++ tests/fixtures/nlp/nlu/with_brand_and_city.json
