+++ tests/fixtures/nlp/nlu/with_cat_bank.json
