+++ tests/fixtures/nlp/nlu/with_cat_city_country.json
