+++ tests/fixtures/nlp/nlu/with_category_and_city.json
