+++ tests/fixtures/nlp/nlu/with_picasso.json
